[RN] Extract header components for reuse
