+++ react/features/base/react/components/native/index.js
@@ -1,5 +1,7 @@
+export { default as BackButton } from './BackButton';
 export { default as Container } from './Container';
 export { default as Header } from './Header';
+export { default as HeaderLabel } from './HeaderLabel';
 export { default as Link } from './Link';
 export { default as LoadingIndicator } from './LoadingIndicator';
 export { default as NavigateSectionListEmptyComponent } from
