+++ react/features/settings/components/native/SettingsView.js
@@ -7,25 +7,24 @@ import {
     SafeAreaView,
     ScrollView,
     Switch,
-    Text,
     TextInput,
     View
 } from 'react-native';
 import { connect } from 'react-redux';
 
 import { translate } from '../../../base/i18n';
-import { Header } from '../../../base/react';
+import { BackButton, Header } from '../../../base/react';
 
 import {
     AbstractSettingsView,
     _mapStateToProps
 } from '../AbstractSettingsView';
 import { setSettingsViewVisible } from '../../actions';
-import BackButton from './BackButton';
 import FormRow from './FormRow';
 import FormSectionHeader from './FormSectionHeader';
 import { normalizeUserInputURL } from '../../functions';
 import styles from './styles';
+import { HeaderLabel } from '../../../base/react/components/native';
 
 /**
  * The native container rendering the app settings page.
@@ -205,13 +204,7 @@ class SettingsView extends AbstractSettingsView {
         return (
             <Header>
                 <BackButton onPress = { this._onRequestClose } />
-                <Text
-                    style = { [
-                        styles.text,
-                        Header.textStyle
-                    ] }>
-                    { this.props.t('settingsView.header') }
-                </Text>
+                <HeaderLabel labelKey = 'settingsView.header' />
             </Header>
         );
     }
